+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix title info was not submitted to Dialog Insight web service
 * Fix capability "aquatonic_course" for shop_manager
 * Fix dropdown menu buttons margins
+* Fix dashboard now values now only use capacity and realtime, not bookings 
 
 ### 1.4.3: May 24th, 2021
 * Times dropdown must be above header in fixed position